Get rid of the handler function.
