+++ packages/unmock-node/src/backend/index.ts
@@ -1,9 +1,40 @@
 import { IncomingMessage, ServerResponse } from "http";
 import Mitm from "mitm";
-import { IBackend, UnmockOptions } from "unmock-core";
+import {
+  FindResponse,
+  IBackend,
+  ISerializedRequest,
+  ISerializedResponse,
+  UnmockOptions,
+} from "unmock-core";
+import { serializeRequest } from "../serialize";
 import * as constants from "./constants";
-import { handleRequestResponse } from "./handler";
 import { responseFinderFactory } from "./response-finder";
+
+const respondFromSerializedResponse = (
+  serializedResponse: ISerializedResponse,
+  res: ServerResponse,
+) => {
+  // TODO Headers
+  res.statusCode = serializedResponse.statusCode;
+  res.write(serializedResponse.body || "");
+  res.end();
+};
+
+export const handleRequestResponse = async (
+  findResponse: FindResponse,
+  req: IncomingMessage,
+  res: ServerResponse,
+) => {
+  const serializedRequest: ISerializedRequest = await serializeRequest(req);
+  const serializedResponse: ISerializedResponse | undefined = findResponse(
+    serializedRequest,
+  );
+  if (!serializedResponse) {
+    throw Error(constants.MESSAGE_FOR_MISSING_MOCK);
+  }
+  respondFromSerializedResponse(serializedResponse, res);
+};
 
 let mitm: any;
 export default class NodeBackend implements IBackend {
